+++ previous-versions/accounting/bankaccounttransactions.go
@@ -353,12 +353,12 @@ func (s *bankAccountTransactions) List(ctx context.Context, request operations.L
 	case httpRes.StatusCode == 200:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out *shared.BankTransactionsResponse
+			var out *shared.BankTransactions
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out); err != nil {
 				return nil, err
 			}
 
-			res.BankTransactionsResponse = out
+			res.BankTransactions = out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}