+++ previous-versions/accounting/RELEASES.md
@@ -361,3 +361,13 @@ Based on:
 - [go v0.1.0] previous-versions/accounting
 ### Releases
 - [Go v0.1.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.1.0 - previous-versions/accounting
+
+## 2023-09-05 00:20:17
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.1.1] previous-versions/accounting
+### Releases
+- [Go v0.1.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.1.1 - previous-versions/accounting